requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-payments==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
fonttools==4.57.0
idna==3.10
phonenumberslite==9.0.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
